build, engine: rename/shuffle builders, buildAndDeployers
